rename selector files
