+++ app/selectors/multichain/multichain.test.ts
@@ -16,7 +16,7 @@ import {
   selectMultichainSelectedAccountCachedBalance,
   selectMultichainShouldShowFiat,
   selectMultichainConversionRate,
-} from './multichainNonEvm';
+} from './multichain';
 import { InternalAccount } from '@metamask/keyring-internal-api';
 import { CHAIN_IDS } from '@metamask/transaction-controller';
 import {
